fix: test was passing but needed some changes to the control of the board, will refactor it ASAP
