+++ tests/equip_test.py
@@ -1,6 +1,6 @@
 import pytest
 from factories.actor_factory import get_new_actor
-from factories.equip_factory import get_new_equip 
+from factories.equip_factory import get_new_equip
 import constants.equips_cons as cons
 
 @pytest.fixture
@@ -10,5 +10,5 @@ def actor():
 def test_zarabatana():
     equip = get_new_equip(**cons.ZARABA)
     actor = get_new_actor(equip=equip)
-    
+
     assert actor.has_command(cons.ZARABA["commands_ids"][0])
